rename kebab-case files - supercedes #2890
