+++ .gitignore
@@ -3,7 +3,7 @@
 .nyc_output
 node_modules
 *.map
-/src/compiler/compile/internal-exports.ts
+/src/compiler/compile/internal_exports.ts
 /compiler.d.ts
 /compiler.*js
 /index.*js
