+++ rollup.config.js
@@ -66,7 +66,7 @@ export default [
 						if (dir === 'internal') {
 							const mod = bundle['index.mjs'];
 							if (mod) {
-								fs.writeFileSync('src/compiler/compile/internal-exports.ts', `// This file is automatically generated\nexport default new Set(${JSON.stringify(mod.exports)});`);
+								fs.writeFileSync('src/compiler/compile/internal_exports.ts', `// This file is automatically generated\nexport default new Set(${JSON.stringify(mod.exports)});`);
 							}
 						}
