+++ src/compiler/compile/render_dom/index.ts
