+++ src/compiler/compile/render_dom/wrappers/AwaitBlock.ts
