+++ src/compiler/compile/render_dom/wrappers/Element/Binding.ts
