+++ src/compiler/compile/render_dom/wrappers/Element/StyleAttribute.ts
