+++ src/compiler/compile/render_dom/wrappers/Window.ts
