+++ src/compiler/compile/render_dom/wrappers/shared/add_event_handlers.ts
