+++ src/compiler/compile/render_dom/wrappers/shared/get_context_merger.ts
