+++ src/compiler/compile/render_ssr/handlers/AwaitBlock.ts
