+++ src/compiler/compile/render_ssr/handlers/Head.ts
