+++ src/compiler/compile/render_ssr/handlers/IfBlock.ts
