+++ src/compiler/compile/render_ssr/handlers/InlineComponent.ts
