+++ src/runtime/internal/index.ts
@@ -1,8 +1,8 @@
 export * from './animations';
-export * from './await-block';
+export * from './await_block';
 export * from './dom';
 export * from './environment';
-export * from './keyed-each';
+export * from './keyed_each';
 export * from './lifecycle';
 export * from './loop';
 export * from './scheduler';
